src/main/NSN/requirements.txt
Flask==3.0.2
Flask-Bcrypt==1.0.1
mysql-connector-python==8.3.0
Werkzeug==3.0.1
pytest==8.0.2
python-dotenv==1.0.1
# mysql-connector-python-rf==2.2.3
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0 